Moved buttons into a folder & update paths
