+++ frontend/src/components/table/TableCell.jsx
@@ -1,4 +1,4 @@
-import { CopyButton } from '../CopyButton';
+import { CopyButton } from '../buttons/CopyButton';
 
 export const TableCell = ({ value }) => {
 	return (
